Introduce a Normal type
Up until now, normal vectors computed by the sphere for the Hit entry
could either point inwards or outwards, which could cause misuse issues
at the call site. Making Normal a strong type ensures that both cases
are handled. The alternative is to always have normal vectors pointing
outwards (the shape inverts the vector if needed), but we then lose the
information when painting and have to compute it again (which is
inconvenient at best).