Render remaining rows locally
This allows the client not to have to spawn a server for local
processing, which is more efficient (at the cost of the added
complexity) and unifies local and remote processing.

The current implementation is copy/conversion-heavy, which could be
avoided by sending floats over the network, but that would have its own
(network) cost as well.